requirements.txt
pyyaml==6.0
discord.py==2.0.1
